+++ src/api/modules/utils.ts
@@ -133,10 +133,6 @@ export const appIdFromRegistry = (app: string, majorLocator: string) => {
     .catch(handleError(app))
 }
 
-const FREE_BILLING_OPTIONS_TYPE = 'free'
-export const isFreeApp = ({ type, free }: { type?: string; free?: boolean }) =>
-  type === FREE_BILLING_OPTIONS_TYPE || free
-
 export async function checkBuilderHubMessage(cliRoute: string): Promise<any> {
   const http = axios.create({
     baseURL: `https://vtex.myvtex.com`,