+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+- App purchases and all code related to it
+
 ## [3.7.3-beta] - 2021-05-31
 
 ### Changed